+++ src/bl64-ans.bash
@@ -56,6 +56,7 @@ function bl64_ans_run_ansible() {
 
   unset ANSIBLE_CONFIG
   unset ANSIBLE_COLLECTIONS
+  bl64_dbg_lib_show_info "extra args: [${debug}]"
   # shellcheck disable=SC2086
   "$BL64_ANS_CMD_ANSIBLE" \
     $debug \
@@ -80,15 +81,15 @@ function bl64_ans_run_ansible_galaxy() {
   bl64_check_parameters_none "$#" || return $?
   local debug=' '
 
-  bl64_check_module_setup "$BL64_ANS_MODULE" &&
-    bl64_check_command "$BL64_ANS_CMD_ANSIBLE_GALAXY" ||
+  bl64_check_module_setup "$BL64_ANS_MODULE" ||
     return $?
 
-  bl64_msg_verbose_lib_enabled && debug="${BL64_ANS_SET_VERBOSE} ${BL64_ANS_SET_DIFF}"
-  bl64_dbg_lib_command_enabled && debug="$BL64_ANS_SET_DEBUG"
+  bl64_msg_verbose_lib_enabled && debug="$BL64_ANS_SET_VERBOSE"
 
   unset ANSIBLE_CONFIG
   unset ANSIBLE_COLLECTIONS
+  bl64_dbg_lib_show_info "extra args: [${debug}]"
+  # shellcheck disable=SC2086
   "$BL64_ANS_CMD_ANSIBLE_GALAXY" \
     $debug \
     "$@"
@@ -112,8 +113,7 @@ function bl64_ans_run_ansible_playbook() {
   bl64_check_parameters_none "$#" || return $?
   local debug=' '
 
-  bl64_check_module_setup "$BL64_ANS_MODULE" &&
-    bl64_check_command "$BL64_ANS_CMD_ANSIBLE_PLAYBOOK" ||
+  bl64_check_module_setup "$BL64_ANS_MODULE" ||
     return $?
 
   bl64_msg_verbose_lib_enabled && debug="${BL64_ANS_SET_VERBOSE} ${BL64_ANS_SET_DIFF}"
@@ -121,6 +121,8 @@ function bl64_ans_run_ansible_playbook() {
 
   unset ANSIBLE_CONFIG
   unset ANSIBLE_COLLECTIONS
+  bl64_dbg_lib_show_info "extra args: [${debug}]"
+  # shellcheck disable=SC2086
   "$BL64_ANS_CMD_ANSIBLE_PLAYBOOK" \
     $debug \
     "$@"