+++ src/bl64-gcp-env.bash
@@ -1,13 +1,13 @@
 #######################################
 # BashLib64 / Module / Globals / Interact with GCP CLI
 #
-# Version: 1.1.0
+# Version: 1.2.0
 #######################################
 
 # Optional module. Not enabled by default
 export BL64_GCP_MODULE="$BL64_LIB_VAR_OFF"
 
-export BL64_GCP_CMD_GCLOUD="${BL64_GCP_CMD_GCLOUD:-}"
+export BL64_GCP_CMD_GCLOUD="$BL64_LIB_UNAVAILABLE"
 
 export BL64_GCP_CONFIGURATION_NAME='bl64_gcp_configuration_private'
 export BL64_GCP_CONFIGURATION_CREATED="$BL64_LIB_VAR_FALSE"