+++ src/bl64-gcp-setup.bash
@@ -1,14 +1,14 @@
 #######################################
 # BashLib64 / Module / Setup / Interact with GCP CLI
 #
-# Version: 1.1.0
+# Version: 1.2.0
 #######################################
 
 #######################################
 # Setup the bashlib64 module
 #
 # Arguments:
-#   None
+#   $1: (optional) Full path to the bin path where gcloud commands are
 # Outputs:
 #   STDOUT: None
 #   STDERR: None
@@ -17,9 +17,15 @@
 #   >0: setup failed
 #######################################
 function bl64_gcp_setup() {
-  bl64_dbg_lib_show_function
+  bl64_dbg_lib_show_function "$@"
+  local gcloud_bin="${1:-${BL64_LIB_DEFAULT}}"
 
-  bl64_gcp_set_command &&
+  if [[ "$gcloud_bin" != "$BL64_LIB_DEFAULT" ]]; then
+    bl64_check_directory "$gcloud_bin" ||
+      return $?
+  fi
+
+  bl64_gcp_set_command "$BL64_GCP_CMD_GCLOUD" &&
     bl64_gcp_set_options &&
     BL64_GCP_MODULE="$BL64_LIB_VAR_ON"
 
@@ -28,8 +34,8 @@ function bl64_gcp_setup() {
 #######################################
 # Identify and normalize commands
 #
+# * If no values are providedprovied, try to detect commands looking for common paths
 # * Commands are exported as variables with full path
-# * The caller function is responsible for checking that the target command is present (installed)
 #
 # Arguments:
 #   None
@@ -40,22 +46,28 @@ function bl64_gcp_setup() {
 #   0: always ok
 #######################################
 function bl64_gcp_set_command() {
-  bl64_dbg_lib_show_function
+  bl64_dbg_lib_show_function "$@"
+  local gcloud_bin="$1"
 
-  # Detect gcloud location
-  if [[ -z "$BL64_GCP_CMD_GCLOUD" ]]; then
-    if [[ -x '/usr/bin/gcloud' ]]; then
-      BL64_GCP_CMD_GCLOUD='/usr/bin/gcloud'
+  if [[ "$gcloud_bin" != "$BL64_LIB_DEFAULT" ]]; then
+    if [[ -x '/home/linuxbrew/.linuxbrew/bin/gcloud' ]]; then
+      gcloud_bin='/home/linuxbrew/.linuxbrew/bin'
     elif [[ -x '/opt/homebrew/bin/gcloud' ]]; then
-      BL64_GCP_CMD_GCLOUD='/opt/homebrew/bin/gcloud'
-    elif [[ -x '/home/linuxbrew/.linuxbrew/bin/gcloud' ]]; then
-      BL64_GCP_CMD_GCLOUD='/home/linuxbrew/.linuxbrew/bin/gcloud'
-    else
-      BL64_GCP_CMD_GCLOUD="$BL64_LIB_UNAVAILABLE"
+      gcloud_bin='/opt/homebrew/bin'
+    elif [[ -x '/usr/local/bin/gcloud' ]]; then
+      gcloud_bin='/usr/local/bin'
+    elif [[ -x '/usr/bin/gcloud' ]]; then
+      gcloud_bin='/usr/bin'
     fi
+  else
+    [[ ! -x "${gcloud_bin}/ansible" ]] && gcloud_bin="$BL64_LIB_DEFAULT"
   fi
 
-  return 0
+  if [[ "$gcloud_bin" != "$BL64_LIB_DEFAULT" ]]; then
+    [[ -x "${gcloud_bin}/gcloud" ]] && BL64_GCP_CMD_GCLOUD="${gcloud_bin}/gcloud"
+  fi
+
+  bl64_dbg_lib_show_vars 'BL64_GCP_CMD_GCLOUD'
 }
 
 #######################################