+++ src/bl64-msg-env.bash
@@ -4,10 +4,21 @@
 # Version: 1.7.0
 #######################################
 
+#
+# Verbosity levels
+#
+# * 0: nothing is showed
+# * 1: application messages only
+# * 2: bashlib64 and application messages
+#
+
 export BL64_MSG_VERBOSE_NONE='0'
 export BL64_MSG_VERBOSE_APP='1'
 export BL64_MSG_VERBOSE_LIB='2'
 
+#
+# Message formats
+#
 export BL64_MSG_FORMAT_PLAIN='R'
 export BL64_MSG_FORMAT_HOST='H'
 export BL64_MSG_FORMAT_TIME='T'
@@ -16,18 +27,18 @@ export BL64_MSG_FORMAT_FULL='F'
 
 export BL64_MSG_FORMAT="${BL64_MSG_FORMAT:-$BL64_MSG_FORMAT_FULL}"
 
-readonly _BL64_MSG_TXT_USAGE='Usage'
-readonly _BL64_MSG_TXT_COMMANDS='Commands'
-readonly _BL64_MSG_TXT_FLAGS='Flags'
-readonly _BL64_MSG_TXT_PARAMETERS='Parameters'
-readonly _BL64_MSG_TXT_ERROR='Error'
-readonly _BL64_MSG_TXT_INFO='Info'
-readonly _BL64_MSG_TXT_TASK='Task'
-readonly _BL64_MSG_TXT_DEBUG='Debug'
-readonly _BL64_MSG_TXT_WARNING='Warning'
-readonly _BL64_MSG_TXT_BATCH='Process'
-readonly _BL64_MSG_TXT_INVALID_FORMAT='invalid format. Please use one of BL64_MSG_FORMAT_*'
-readonly _BL64_MSG_TXT_BATCH_START='started'
-readonly _BL64_MSG_TXT_BATCH_FINISH_OK='finished successfully'
-readonly _BL64_MSG_TXT_BATCH_FINISH_ERROR='finished with errors'
+declare _BL64_MSG_TXT_USAGE='Usage'
+declare _BL64_MSG_TXT_COMMANDS='Commands'
+declare _BL64_MSG_TXT_FLAGS='Flags'
+declare _BL64_MSG_TXT_PARAMETERS='Parameters'
+declare _BL64_MSG_TXT_ERROR='Error'
+declare _BL64_MSG_TXT_INFO='Info'
+declare _BL64_MSG_TXT_TASK='Task'
+declare _BL64_MSG_TXT_DEBUG='Debug'
+declare _BL64_MSG_TXT_WARNING='Warning'
+declare _BL64_MSG_TXT_BATCH='Process'
+declare _BL64_MSG_TXT_INVALID_FORMAT='invalid format. Please use one of BL64_MSG_FORMAT_*'
+declare _BL64_MSG_TXT_BATCH_START='started'
+declare _BL64_MSG_TXT_BATCH_FINISH_OK='finished successfully'
+declare _BL64_MSG_TXT_BATCH_FINISH_ERROR='finished with errors'