+++ src/bl64-py-env.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Globals / Interact with system-wide Python
 #
-# Version: 1.7.0
+# Version: 1.9.0
 #######################################
 
 # Optional module. Not enabled by default
@@ -16,6 +16,9 @@ export BL64_PY_CMD_PYTHON38="$BL64_LIB_UNAVAILABLE"
 export BL64_PY_CMD_PYTHON39="$BL64_LIB_UNAVAILABLE"
 export BL64_PY_CMD_PYTHON310="$BL64_LIB_UNAVAILABLE"
 
+# Full path to the python venv activated by bl64_py_setup
+export BL64_PY_VENV_PATH=''
+
 # Version info
 export BL64_PY_VERSION_PYTHON3=''
 export BL64_PY_VERSION_PIP3=''
@@ -26,14 +29,18 @@ export BL64_PY_SET_PIP_UPGRADE=''
 export BL64_PY_SET_PIP_USER=''
 export BL64_PY_SET_PIP_DEBUG=''
 export BL64_PY_SET_PIP_QUIET=''
+export BL64_PY_SET_PIP_SITE=''
 export BL64_PY_SET_PIP_NO_WARN_SCRIPT=''
 export BL64_PY_SET_VENV_CFG=''
 export BL64_PY_SET_MODULE_VENV=''
 export BL64_PY_SET_MODULE_PIP=''
 
-readonly _BL64_PY_TXT_PIP_PREPARE_PIP='upgrade pip module'
-readonly _BL64_PY_TXT_PIP_PREPARE_SETUP='install and upgrade setuptools modules'
-readonly _BL64_PY_TXT_PIP_INSTALL='install modules'
+declare _BL64_PY_TXT_PIP_PREPARE_PIP='upgrade pip module'
+declare _BL64_PY_TXT_PIP_PREPARE_SETUP='install and upgrade setuptools modules'
+declare _BL64_PY_TXT_PIP_INSTALL='install modules'
+declare _BL64_PY_TXT_VENV_MISSING='requested python virtual environment is missing'
+declare _BL64_PY_TXT_VENV_INVALID='requested python virtual environment is invalid (no pyvenv.cfg found)'
+declare _BL64_PY_TXT_VENV_CREATE='create python virtual environment'
 
 # External commands variables
 export PYTHONHOME